* move handling of selection state into MoveHistoryService
